Move desktop-specific game host logic to `osu.Framework.Desktop`
