+++ osu.Framework.Templates/templates/template-flappy/FlappyDon.Desktop/Program.cs
@@ -1,5 +1,5 @@
 ﻿using FlappyDon.Game;
-using osu.Framework;
+using osu.Framework.Desktop;
 using osu.Framework.Platform;
 
 namespace FlappyDon.Desktop
@@ -8,7 +8,7 @@ namespace FlappyDon.Desktop
     {
         public static void Main()
         {
-            using (GameHost host = Host.GetSuitableDesktopHost(@"FlappyDon"))
+            using (GameHost host = DesktopHost.GetSuitableHost(@"FlappyDon"))
             using (osu.Framework.Game game = new FlappyDonGame())
             {
                 host.Run(game);
