+++ osu.Framework.Templates/templates/template-flappy/FlappyDon.Game.Tests/Program.cs
@@ -1,5 +1,5 @@
 ﻿using FlappyDon.Game.Testing;
-using osu.Framework;
+using osu.Framework.Desktop;
 using osu.Framework.Platform;
 
 namespace FlappyDon.Game.Tests
@@ -8,7 +8,7 @@ namespace FlappyDon.Game.Tests
     {
         public static void Main()
         {
-            using (GameHost host = Host.GetSuitableDesktopHost("visual-tests"))
+            using (GameHost host = DesktopHost.GetSuitableHost("visual-tests"))
             using (var game = new FlappyDonTestBrowser())
                 host.Run(game);
         }
